stm32mp1: add a delay loop early in SPL for debugger attach
Current Soc revision has a known limitation regarding debug linked
to reset halt. For this purpose we couldn't attach inside the rom code.
We may integrate a delay loop early in the SPL boot process to wait
for the debugger to attach.

Today, we use a backup register bit to know if we are in debug mode and
add a loop at the beginning of SPL execution, in the weak function
save_boot_params() called by reset function in arch/cpu/armv7/start.S.

This can be removed when using the Soc revision that fixes the limitation.

Anyway, this source code identifies the Soc revision and is
only executed if it corresponds, so it can be kept on other revisions
without any consequence.

Change-Id: Idd0ff0d790010e29bb6f644b8923c802d191fc1d
Depends-On: I459ee5bd294660ec6cc48cd8ea08ba35be0c51ba
Signed-off-by: Patrick Delaunay <patrick.delaunay@st.com>
Reviewed-on: https://gerrit.st.com/116178
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Patrice CHOTARD <patrice.chotard@st.com>
Reviewed-on: https://gerrit.st.com/116368