fdt: Allow indicating a node is for U-Boot proper only
This add missing parts for previous commit 06f94461a9f4
("fdt: Allow indicating a node is for U-Boot proper only")


At present it is not possible to specify that a node should be used before
relocation (in U-Boot proper) without it also ending up in SPL and TPL
device trees. Add a new "u-boot,dm-pre-proper" boolean property for this.

Tested with "u-boot,dm-pre-proper" for video driver, without live tree.

Change-Id: I456de145f4689539dfa0c9fcda8ff0b6be0238a7
Signed-off-by: Patrick Delaunay <patrick.delaunay@st.com>
Reviewed-on: https://gerrit.st.com/120645
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>