spi: stm32_qspi: Remove STM32_QSPI_CR_FTHRES_MASK
On STM32 F4/F7/H7 SoCs, FTHRES is a 5 bits field in QSPI_CR register,
but for STM32MP1 SoCs, FTHRES is a 4 bits field long.
STM32_QSPI_CR_FTHRES_MASK definition is not correct.

Even if it is currently not used, remove STM32_QSPI_CR_FTHRES_MASK
to avoid confusion in the future.

Change-Id: I3afebd86dc2f413bf013715d712e0e02b93f09f5
Signed-off-by: Patrice Chotard <patrice.chotard@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/134149
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Christophe KERELLO <christophe.kerello@st.com>
Reviewed-by: Patrick DELAUNAY <patrick.delaunay@st.com>