board: stm32mp1: fixup the USB product id for fastboot
Select a specific product id used during USB enumeration
when the fastboot gadget is used.

For ST Microelectronics board we are using
the VID:PID = 0x0483:0x0AFB for fastboot
as defined in http://www.linux-usb.org/usb.ids

This patch avoid conflict when the same USB VID/PID is used for
fastboot and DFU (two different protocols associated to the same PID).

Signed-off-by: Patrick Delaunay <patrick.delaunay@st.com>
Change-Id: Id25154105bef11c7ed8031aed96e7c67b1805885
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/142124
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Christophe GUIBOUT <christophe.guibout@st.com>
Reviewed-by: Patrice CHOTARD <patrice.chotard@st.com>
Reviewed-by: Nicolas LOUBOUTIN <nicolas.louboutin@st.com>
Tested-by: Christophe GUIBOUT <christophe.guibout@st.com>