i2c: stm32f7_i2c: allows for any bus frequency
Do not limitate to 3 (100KHz, 400KHz, 1MHz) bus frequencies, but
instead allow for any frequency. Depending on the requested
frequency (via the clock-frequency DT entry), use the spec
data from either Standard, Fast or Fast Plus mode.

In order to do so, the driver do not use anymore spec identifier
by directly handle the requested frequency and from it retrieve
the corresponding spec data to be used for the computation
of the timing register.

Change-Id: I44e04e48643f0d3554334318ff2111860ca0f8ab
Signed-off-by: Alain Volmat <alain.volmat@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/u-boot/+/144006
Reviewed-by: Patrick DELAUNAY <patrick.delaunay@st.com>